fix: use roundf instead of round

+++ Module02/ex01/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/22 14:07:09 by gwolf             #+#    #+#             */
-/*   Updated: 2023/09/24 08:28:47 by gwolf            ###   ########.fr       */
+/*   Updated: 2023/09/25 20:31:03 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ Fixed::Fixed(int const num) : m_value(num << s_fract_bits)
 	std::cout << "Int constructor called: Fixed\n";
 }
 
-Fixed::Fixed(float const num) : m_value(round(num * (1 << s_fract_bits)))
+Fixed::Fixed(float const num) : m_value(roundf(num * (1 << s_fract_bits)))
 {
 	std::cout << "Float constructor called: Fixed\n";
 }
